### 2024-04-30 #6706 @editorialbot
[PRE REVIEW]: SAnTex: A Python-based Library for Seismic Anisotropy Calculation
**Submitting author:** @utpal-singh (<a href="http://orcid.org/0000-0001-8304-5615">Utpal Singh</a>)
**Repository:** https://github.com/utpal-singh/SAnTex
**Branch with paper.md** (empty if default branch): main
**Version:** v1.2
