### 2024-05-03 #6706 @kthyng
@utpal-singh some comments/questions:
* I see a readthedocs file and docs files, notebooks, example files, but no readthedocs link?
* how should a user verify that the code is working correctly once they install it?
