+++ santex/anisotropy/__init__.py
@@ -1,2 +1,1 @@
 from .anisotropy import Anisotropy
-from .plot_seismic_velocity import plot_vp_2d
