+++ santex/__init__.py
@@ -2,5 +2,4 @@ from .material import Material
 from .ebsd import EBSD
 from .tensor import Tensor
 from .anisotropy import Anisotropy
-from .anisotropy.plot_seismic_velocity import plot_vp_2d
 from .isotropy import Isotropy
