### 2024-07-02 #6886 @rkurchin
If you read the [opening comment](https://github.com/openjournals/joss-reviews/issues/6886#issue-2352067419) on this review issue, it has some basic instructions. Please don't hesitate to let me know what questions you have!

The general flow will be for you to generate your checklist and then work your way through it as you try out the code. Feel free to leave comments here and/or to file issues/PR's in the project repository related to the checklist items. If you do the latter, please link to this review issue for easy tracking (that way I won't have to hound for updates here if I can easily see that there's conversation going on in an issue/PR elsewhere 😉).
