### 2024-07-26 utpal-singh/SAnTex#6 @elena-pascal
JOSS REVIEW -- Documentation is hard to access
Hello @utpal-singh,

This relates to https://github.com/openjournals/joss-reviews/issues/6886

You have so many documentation resources (docs, examples, READMEs) but they are all relatively hidden.
Here are some suggestions:

- [ ] The docstrings are really good but they don't show up in the readthedocs.

- [ ] It is not obvious to me why there is a  README.pdf and a README.md, and the one harder to access has more information.

- [ ] You have a great example folder. It would be great if the README.md would contain a bit of info about them and very basic how to run them.
