### 2024-07-26 #6886 @elena-pascal
Hello, the paper.md seems to have latex maths (md requires extra $ delimiters)? At least on my github it does not render. Can we have access to a copy with rendered maths? pdf would be fantastic.
