### 2024-07-26 #6886 @rkurchin
Yes, the rendered one is accessible in the editorialbot comment above (and can be regenerated by running `@editorialbot generate pdf`): https://github.com/openjournals/joss-reviews/issues/6886#issuecomment-2166798638
