### 2024-08-22 #6886 @elena-pascal
Hi @rkurchin, I made an issue for the insufficient documentation, it would be easier for me to proceed with assessing the software if I know how to use it.
