### 2024-09-03 #6886 @utpal-singh
Hi @elena-pascal,

Thank you for reviewing the paper. I wanted to confirm if the documentation you're requesting should be in the readthedocs format? Currently, we have it in PDF format, but I can create the readthedocs version if that's what's needed.

Best regards,
Utpal
