### 2024-09-03 #6886 @elena-pascal
Hi @utpal-singh,

I left you [an issue with more details](https://github.com/utpal-singh/SAnTex/issues/6).
Is there a reason you choose your documentation to be in a pdf file? Git doesn't really deal well with binary files, additionally pdf is not a great way to store code snippets.
