### 2024-09-03 #6886 @utpal-singh
Hi @elena-pascal

Thanks for the elaborated review issue. It seems I had missed the elaborated issue. Thanks for the feedback. The reason for providing the pdf file was a relatively quick look at the package contents. But, I will also prepare the documentation in the readthedocs too, not really much of an issue.
