### 2024-09-04 cb1bd52
readme docs html entry point
