+++ docs/generated/santex.rst
@@ -1,0 +1,23 @@
+﻿santex
+======
+
+.. automodule:: santex
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
