+++ docs/modules.rst
@@ -1,8 +1,0 @@
-SAnTex
-======
-
-.. toctree::
-   :maxdepth: 4
-
-   santex
-   setup
