+++ docs/generated/santex.rst
@@ -1,23 +1,0 @@
-﻿santex
-======
-
-.. automodule:: santex
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   
-
-   
-   
-   
-
-
-
