+++ docs/api.rst
@@ -1,7 +1,0 @@
-API
-===
-
-.. autosummary::
-   :toctree: generated
-
-   santex
