+++ docs/index.rst
@@ -1,22 +1,0 @@
-.. santex documentation master file, created by
-   sphinx-quickstart on Thu Apr 18 21:49:28 2024.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to santex's documentation!
-==================================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-   modules
-   usage
-   api
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
