+++ santex/anisotropy/anisotropy.py
@@ -20,45 +20,53 @@ import math
 
 class Anisotropy:
     """
-    Anisotropy class for representing anisotropic material properties.
+    A class to represent anisotropic material properties.
 
     Attributes:
-        cijkl (numpy.ndarray or None): stiffness tensor in voigt notation.
-            It is converted from the input stiffness matrix using voigt_to_tensor method.
-            If stiffness_matrix is None, cijkl remains None. Units can be SI or CGS but should remain consistent
-            between stiffness matrix and density
-        rho (float or None): Density of the material.
-            If density is None, rho remains None.
+        cijkl (Optional[numpy.ndarray]): Stiffness tensor in Voigt notation.
+            It is converted from the input stiffness matrix using the `voigt_to_tensor` method.
+            If `stiffness_matrix` is None, `cijkl` remains None. Units should be consistent between the 
+            stiffness matrix and density, either in SI or CGS.
+        rho (Optional[float]): Density of the material.
+            If `density` is None, `rho` remains None.
+
+    Parameters:
+        stiffness_matrix (Optional[numpy.ndarray]): A 6x6 matrix representing the stiffness tensor in Voigt notation.
+        density (Optional[float]): The density of the material.
 
     Methods:
         __init__(stiffness_matrix, density):
             Initializes the Anisotropy object with the given stiffness_matrix and density.
-            If either stiffness_matrix or density is None, cijkl and rho remain None.
+            If either `stiffness_matrix` or `density` is None, `cijkl` and `rho` remain None.
 
-    Example Usage:
-        stiffness_matrix = np.array([[198.96, 73.595, 68.185, 0., 9.735, 0.],
-                                [73.595, 155.94, 62.23, 0., 6.295, 0.],
-                                [68.185, 62.23, 225.99, 0., 33.85, 0.],
-                                [0., 0., 0., 65.66, 0., 6.415],
-                                [9.735, 6.295, 33.85, 0., 60.23, 0.],
-                                [0., 0., 0., 6.415, 0., 65.18]]) * 10**9
-        density = 3.5 
-        anisotropy = Anisotropy(stiffness_matrix, density)
+    Example:
+        >>> import numpy as np
+        >>> stiffness_matrix = np.array([[198.96, 73.595, 68.185, 0., 9.735, 0.],
+        >>>                              [73.595, 155.94, 62.23, 0., 6.295, 0.],
+        >>>                              [68.185, 62.23, 225.99, 0., 33.85, 0.],
+        >>>                              [0., 0., 0., 65.66, 0., 6.415],
+        >>>                              [9.735, 6.295, 33.85, 0., 60.23, 0.],
+        >>>                              [0., 0., 0., 6.415, 0., 65.18]]) * 10**9
+        >>> density = 3.5  # density in g/cm³ or consistent units
+        >>> anisotropy = Anisotropy(stiffness_matrix, density)
+        >>> print(anisotropy.cijkl)
+        >>> print(anisotropy.rho)
     """
+
     
     def __init__(self, stiffness_matrix, density):
         """
-        Initializes an Anisotropy object with the given stiffness_matrix and density.
+        Initializes an Anisotropy object with the given stiffness matrix and density.
 
         Parameters:
-            stiffness_matrix (list of lists or numpy.ndarray): The 6x6 stiffness matrix representing
+            stiffness_matrix (list[list[float]] or numpy.ndarray): A 6x6 stiffness matrix representing
                 the material's anisotropic properties in Voigt notation.
             density (float): The density of the material.
 
         Raises:
-            TypeError: If stiffness_matrix is not a 6x6 list or numpy.ndarray.
-
[…]
