+++ docs/source/santex.ebsd.rst
@@ -1,0 +1,61 @@
+santex.ebsd package
+===================
+
+Submodules
+----------
+
+santex.ebsd.calcGrainBoundary module
+------------------------------------
+
+.. automodule:: santex.ebsd.calcGrainBoundary
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+santex.ebsd.ebsd module
+-----------------------
+
+.. automodule:: santex.ebsd.ebsd
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+santex.ebsd.melt\_tensor module
+-------------------------------
+
+.. automodule:: santex.ebsd.melt_tensor
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+santex.ebsd.odf module
+----------------------
+
+.. automodule:: santex.ebsd.odf
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+santex.ebsd.rotateEBSD module
+-----------------------------
+
+.. automodule:: santex.ebsd.rotateEBSD
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+santex.ebsd.saveEBSD module
+---------------------------
+
+.. automodule:: santex.ebsd.saveEBSD
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: santex.ebsd
+   :members:
+   :undoc-members:
+   :show-inheritance:
