+++ docs/source/index.rst
@@ -1,0 +1,28 @@
+.. santex documentation master file, created by
+   sphinx-quickstart on Wed Sep  4 16:31:06 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to santex's documentation!
+==================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   modules.rst
+   santex.anisotropy.rst
+   santex.ebsd.rst
+   santex.isotropy.rst
+   santex.material.rst
+   santex.rst
+   santex.tensor.rst
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
