+++ docs/source/santex.material.rst
@@ -1,0 +1,21 @@
+santex.material package
+=======================
+
+Submodules
+----------
+
+santex.material.material module
+-------------------------------
+
+.. automodule:: santex.material.material
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: santex.material
+   :members:
+   :undoc-members:
+   :show-inheritance:
