+++ docs/source/santex.anisotropy.rst
@@ -1,0 +1,77 @@
+santex.anisotropy package
+=========================
+
+Submodules
+----------
+
+santex.anisotropy.anisotropy module
+-----------------------------------
+
+.. automodule:: santex.anisotropy.anisotropy
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+santex.anisotropy.map module
+----------------------------
+
+.. automodule:: santex.anisotropy.map
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+santex.anisotropy.melt module
+-----------------------------
+
+.. automodule:: santex.anisotropy.melt
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+santex.anisotropy.modalAnisotropy module
+----------------------------------------
+
+.. automodule:: santex.anisotropy.modalAnisotropy
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+santex.anisotropy.plot\_seismic\_velocity module
+------------------------------------------------
+
+.. automodule:: santex.anisotropy.plot_seismic_velocity
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+santex.anisotropy.plot\_vel\_grid module
+----------------------------------------
+
+.. automodule:: santex.anisotropy.plot_vel_grid
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+santex.anisotropy.utils module
+------------------------------
+
+.. automodule:: santex.anisotropy.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+santex.anisotropy.vtkplotter module
+-----------------------------------
+
+.. automodule:: santex.anisotropy.vtkplotter
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: santex.anisotropy
+   :members:
+   :undoc-members:
+   :show-inheritance:
