+++ docs/requirements.txt
@@ -1,0 +1,6 @@
+sphinx==6.2.1
+sphinx-rtd-theme==1.2.2
+# pillow==5.4.1
+mock==1.0.1
+commonmark==0.9.1
+recommonmark==0.5.0
