### 2024-09-04 284a6d4
updated conf for docs
