+++ docs/requirements.txt
@@ -4,3 +4,4 @@ sphinx-rtd-theme==1.2.2
 mock==1.0.1
 commonmark==0.9.1
 recommonmark==0.5.0
+sphinx_book_theme
