+++ .readthedocs.yaml
@@ -12,5 +12,5 @@ python:
      - requirements: docs/requirements.txt
 
 sphinx:
-  configuration: docs/conf.py
+  configuration: docs/source/conf.py
 
