+++ docs/requirements.txt
@@ -5,3 +5,6 @@ mock==1.0.1
 commonmark==0.9.1
 recommonmark==0.5.0
 sphinx_book_theme
+sphinx
+sphinx-autodoc
+readthedocs-sphinx-ext
