+++ docs/source/conf.py
@@ -30,6 +30,7 @@ extensions = [
 templates_path = ['_templates']
 exclude_patterns = []
 
+autosummary_generate = True
 
 
 # -- Options for HTML output -------------------------------------------------
