### 2024-09-04 4a48616
updated conf for docs
