### 2024-09-05 3f88d5b
Update .readthedocs.yaml
