+++ docs/source/install.rst
@@ -1,0 +1,17 @@
+Santex can be installed using:
+
+.. code:: bash
+
+   pip install santex
+
+To build from source:
+-----------------------
+
+Open the terminal and go to Downloads or Documents directory by typing cd
+~/Documents or cd ~/Downloads or any other directory, where you can store the
+repository of SAnTex. Run the following to get the updated version of SAnTex
+
+.. code:: bash
+    git clone https://github.com/utpal-singh/SAnTex.git
+    cd santex
+    pip install .
