+++ docs/source/workingOnSantex.rst
@@ -1,0 +1,222 @@
+Working with SAnTex
+========================
+
+In a Python working environment, import SAnTex's modules and Python libraries:
+
+.. code-block:: python
+
+    from santex.ebsd import EBSD
+    from santex.tensor import Tensor
+    from santex.material import Material
+    from santex.isotropy import Isotropy
+    import pandas as pd
+
+
+Load the EBSD file:
+
+.. code-block:: python
+
+    ebsdfile = EBSD("ebsd.ctf")
+
+
+This loads the `ebsd.ctf` file in the Python object `ebsdfile` of class `EBSD`, from which users are able to call methods for further processing, cleaning of the ebsd data, and ultimately calculate elastic properties.
+
+
+This loads the ebsd.ctf file in the python object ebsdfile of class EBSD, from which
+users are able to call methods for further processing, cleaning of the ebsd data, and
+ultimately calculate elastic properties.
+
+*Class:* In Python, a class is a blueprint for creating objects. It defines the structure
+and behavior of objects of that type. Think of it as a template or a prototype that
+encapsulates data (attributes) and methods (functions) that operate on that data.
+
+*Object:* An object, also known as an instance, is a unique entity that is created based
+on the structure defined by its class. When an object is instantiated, it inherits the
+attributes and methods defined in its class, but can also have its own unique state
+and behavior.
+
+**Note:** The `ebsd.ctf` should be located in the directory this notebook is being run from. But you may provide a relative path like ``data/ebsd.ctf`` or an absolute import such as ``/Users/myname/Documents/ebsd/ebsd.ctf``.
+
+
+Available methods in the EBSD class
+==========
+
+To list the phases present in the ebsd file:
+
+
+.. code-block:: python
+
+    phases_available = ebsdfile.phases()
+    print(phases_available)
+
+Load the ebsd file into a pandas dataframe ``df``, via invoking the ``get_ebsd_data()``
+method from the EBSD class: *Note:* this dataframe ``df`` is critical as it holds
+information of all the euler angles and the ebsd data.
+
+
+.. code-block:: python
+
+    df = ebsdfile.get_ebsd_data()
+
+To look into the contents of the `df`, enter the following directive:
+
+.. code-block:: python
+
+    print(df)
+
+To look into the data header of the EBSD file, following can be used:
+
+.. code-block:: python
+
+    ebsd_data_header = ebsdfile.get_ebsd_data_header()
+    print(ebsd_data_header)
+
+To look at the index of phase, phases, and relative abundence, follwoing can be entered:
+
+.. code-block:: python
+
+    phases = ebsdfile.phases()
+    print(phases)
+
+To get the euler angles of a certain phase, lets say phase = 1 in this case is Forsterite, enter the following:
+
+.. code-block:: python
+
+    forsteriteEulerAngles = ebsdfile.get_euler_angles(phase = 1, data = df)
+    print(forsteriteEulerAngles)
+
+To plot this ebsd file, following can be enterred:
+
+.. code-block:: python
+
+    ebsdfile.plot(df)
+
+To save this image, following can be used:
+
+.. code-block:: python
+
+    ebsdfile.plot(df, cmap = "viridis", save_image= True, image_filename= "ebsd.jpg", legend_location="lower right")
+
+
+Plotting Conventions
+===================
+
+Following are the keywords to orient sample reference plane and to store in a new
+dataframe, and plot it
+
+.. code-block:: bash
[…]
