### 2024-09-05 6e256ed
updated methods on ebsd in docs and anisotropy
