+++ docs/source/conf.py
@@ -24,6 +24,7 @@ extensions = [
    'sphinx.ext.doctest',
    'sphinx.ext.autodoc',
    'sphinx.ext.autosummary',
+   'sphinx.ext.mathjax',
 ]
 
 
