+++ docs/source/workingOnSantex.rst
@@ -215,8 +215,217 @@ Compare original and clean datasets
 
 
 .. code-block:: python
-    
+
     ebsdfile.plot()
     ebsdfile.plot(df = filtered_df_grain_boundary)
 
 
+Calculating Anisotropy from EBSD file:
+==============
+
+To prepare the dataframe, the user can instanciate Material class as follows:
+
+.. code-block:: python
+
+    material_instance = Material()
+
+The densities are within the santex package registry, and can be accessed via the following directive. The pressure in GPa and Temperature can be entered as:
+
+.. code-block:: python
+
+    rho_Fo = material_instance.load_density("Forsterite", 2, 1500)
+    rho_diop = material_instance.load_density("Diopside", 2, 1500)
+    rho_ens = material_instance.load_density("Enstatite", 2, 1500)
+
+The registry for elastic tensors within santex can be accessed via the following commands. The pressure and temperacture
+conditions can be entered via the keyword PRESSURE and TEMP
+
+.. code-block:: python
+
+    cij_Fo = material_instance.voigthighPT('Forsterite', PRESSURE = 2, TEMP = 1500)
+    cij_ens = material_instance.voigthighPT('Enstatite', PRESSURE = 2, TEMP = 1500)
+    cij_diop = material_instance.voigthighPT('Diopside', PRESSURE = 2, TEMP = 1500)
+
+For preparing dataframes for seismic anisotropy, the elasticstiffness tensors and densities needs to be passed as a list, as:
+
+.. code-block:: python
+
+    cij = [cij_Fo, cij_ens, cij_diop]
+    density = [rho_Fo, rho_ens, rho_diop]
+
+The euler angles for the phases should also be bundled in list as:
+
+.. code-block:: python
+
+    forsterite = ebsdfile.get_euler_angles(phase = 1, data=filtered_df_grain_boundary)
+    enstatite = ebsdfile.get_euler_angles(phase = 2, data=filtered_df_grain_boundary)
+    diopside = ebsdfile.get_euler_angles(phase = 3, data=filtered_df_grain_boundary)
+    euler_angles = [forsterite, enstatite, diopside]
+
+The anisotropy can then be calculated as following:
+
+.. code-block:: python
+
+    average_tensor, average_density = ebsdfile.getAnisotropyForEBSD(cij, euler_angles, density)
+    anis = Anisotropy(average_tensor*10**9, average_density)
+
+To look at the plots for the seismic velocities, following command can be entered:
+
+.. code-block:: python
+
+    anis.plot()
+
+To look at the anisotropy values, maxvp, minvp, maxvs1, minvs1, etc.., folllowing command can be used:
+
+.. code-block:: python
+
+    anis.anisotropy_values()
+
+This gives values such as:
+
+.. code-block:: bash
+
+    Max Vp:  8449.545069720665
+    Min Vp:  7834.1684457727015
+    Max Vs1:  4843.772824149944
+    Min Vs1:  4577.275239228743
+    Max Vs2:  4661.180158160457
+    Min Vs2:  4555.315881762446
+    Max vs anisotropy percent:  6.110766862825683
+    Min vs anisotropy percent:  0.020842429078130564
+    P wave anisotropy percent:  7.558185340984474
+    S1 Wave anisotropy percent:  5.657493372889703
+    S2 Wave anisotropy percent:  2.297278183366879
+    Velocity difference:  287.209223367252
+    Vp/Vs1 ratio:  3.4012758430117644
+
+
+Ordinary Distribution Function (ODF)
+=========
+
+The orientation distribution function (ODF) is a function on the orientation space that associates to each orientation g the volume percentage of crystals in a polycrystaline specimen that are in this specific orientation
+
+ODF can be calculated and plotted as:
+
+.. code-block:: python
+
[…]
