+++ docs/requirements.txt
@@ -1,1 +1,2 @@
 sphinx_book_theme
+sphinx.ext.mathjax
