### 2024-09-05 1350dee
corrected requirements.txt for sphinx
