### 2024-09-09 3ab9b59
docs--> added method and modal mineral composition
