+++ docs/source/workingOnSantex.rst
@@ -429,3 +429,150 @@ In order to illustrate the concept of inverse pole figures, let's calculate the 
         vector_title='Z',
         projection='ipf',
         crystal_symmetry='D2',)
+
+
+Material analysis
+===================
+
+The Material class from the Sage library is used for defining and working with materials calculations.
+
+Material Module can be loaded as:
+
+.. code-block:: python
+
+    import numpy as np
+    from tabulate import tabulate
+    from santex.material import Material
+
+The available material within the santex registry can be viewed as:
+
+.. code-block:: python
+
+    material_instance = Material()
+    phases_info = material_instance.availablePhases()
+    print("Available Phases:")
+    print(phases_info)
+
+This loads a list of materials present with their crystal systems and primary phase information as shown below:
+
+.. code-block:: bash
+
+    Available Phases:
+    +---------------------------------------+---------------------+-----------------------------------+
+    |                 Phase                 |   Crystal System    |           Primary Phase           |
+    +---------------------------------------+---------------------+-----------------------------------+
+    |           Almandine-pyrope            |        Cubic        |              Garnet               |
+    |               Grossular               |        Cubic        |              Garnet               |
+    |               Majorite                |        Cubic        |              Garnet               |
+    |                Pyrope                 |        Cubic        |              Garnet               |
+    |              a_quartz_1               | Hexagonal/ Trigonal |              Quartz               |
+    |              a_quartz_2               | Hexagonal/ Trigonal |              Quartz               |
+    |              a_Quartz_3               | Hexagonal/ Trigonal |              Quartz               |
+    |              a_quartz_4               | Hexagonal/ Trigonal |              Quartz               |
+    |             a_quartz_696C             | Hexagonal/ Trigonal |              Quartz               |
+    |             a_quartz_700C             | Hexagonal/ Trigonal |              Quartz               |
+    |               Calcite_1               | Hexagonal/ Trigonal |              Calcite              |
+    |               Calcite_2               | Hexagonal/ Trigonal |              Calcite              |
+    |              Forsterite               |    Orthorhombic     |              Olivine              |
+    |               Fayalite                |    Orthorhombic     |              Olivine              |
+    |               Lawsonite               |    Orthorhombic     |              Olivine              |
+    |        Orthoenstatite (MgSiO3)        |    Orthorhombic     |              Olivine              |
+    |        Orthoenstatite (MgSiO3)        |    Orthorhombic     |              Olivine              |
+    |               Enstatite               |    Orthorhombic     |              Olivine              |
+    |       Bronzite (Mg0.8Fe0.2SiO3)       |    Orthorhombic     |              Olivine              |
+    |         Ferrosilite (FeSiO3)          |    Orthorhombic     |              Olivine              |
+    |                Biotite                |     Monoclinic      | Phyllosilicates and clay minerals |
+    |               Muscovite               |     Monoclinic      | Phyllosilicates and clay minerals |
+    |              Phlogopite               |     Monoclinic      | Phyllosilicates and clay minerals |
+    |            Illite-smectite            |     Monoclinic      | Phyllosilicates and clay minerals |
+    |                Dickite                |     Monoclinic      | Phyllosilicates and clay minerals |
[…]
