### 2024-09-23 85115a0
added some docs and modified anisotropy.py
