+++ docs/source/workingOnSantex.rst
@@ -576,3 +576,398 @@ To plot the seismic velocities:
     anisotropy.plot()
 
 
+Isotropic Velocities analysis
+----------------
+
+Isotropy module can be loaded as follows:
+
+.. code-block:: python
+
+    from santex.isotropy import Isotropy
+    isotropy = Isotropy()
+
+To check the available phases, a user can invoke ``get_available_phases()`` within the Isotropy class as
+
+.. code-block:: python
+
+    isotropy.get_available_phases()
+
+This gives a list of phases which are available in santex registry as is shown in below example
+
+.. code-block:: bash
+
+    ######################Available Phases######################
+    Material id: aqz       Material name: Alpha-Quartz
+    ######################Available Phases######################
+    Material id: bqz       Material name: Beta-Quartz
+    ######################Available Phases######################
+    Material id: coe       Material name: Coesite
+    ######################Available Phases######################
+    Material id: hAb       Material name: High-T Albite
+    ######################Available Phases######################
+    Material id: lAb       Material name: Low-T Albite
+    ######################Available Phases######################
+    Material id: an       Material name: Anorthite
+    ######################Available Phases######################
+    Material id: or       Material name: Orthoclase
+    ######################Available Phases######################
+    Material id: san       Material name: Sanidine
+    ######################Available Phases######################
+    Material id: alm       Material name: Almandine
+    ######################Available Phases######################
+    Material id: gr       Material name: Grossular
+    ######################Available Phases######################
+    Material id: py       Material name: Pyrope
+    ######################Available Phases######################
+    Material id: fo       Material name: Forsterite
+    ######################Available Phases######################
+    Material id: fa       Material name: Fayalite
+    ######################Available Phases######################
+    Material id: en       Material name: Enstatite
+    ######################Available Phases######################
+    Material id: fs       Material name: Ferrosillite
+    ######################Available Phases######################
+    Material id: mgts       Material name: Mg-Tschermak
+    ######################Available Phases######################
+    Material id: di       Material name: Diopside
+    ######################Available Phases######################
+    Material id: hed       Material name: Hedenbergite
+    ######################Available Phases######################
+    Material id: jd       Material name: Jadeite
+    ######################Available Phases######################
+    Material id: ac       Material name: Acmite
+    ######################Available Phases######################
+    Material id: cats       Material name: Ca-Tschermak
+    ######################Available Phases######################
+    Material id: gl       Material name: Glaucophane
+    ######################Available Phases######################
+    Material id: fgl       Material name: Ferroglaucophane
+    ######################Available Phases######################
+    Material id: tr       Material name: Tremolite
+    ######################Available Phases######################
+    Material id: fact       Material name: Ferroactinolite
+    ######################Available Phases######################
+    Material id: ts       Material name: Tschermakite
[…]
