+++ docs/source/workingOnSantex.rst
@@ -427,6 +427,204 @@ In order to illustrate the concept of inverse pole figures, let's calculate the 
         crystal_symmetry='D2',)
 
 
+Tensor analysis
+===========
+
+Stiffness tensors of phases are a 3*3*3*3 size, however, due to its symmetricity, it can also be written as a voigt notation of 6*6 matrix. A tensor can be rotated
+in Bunge Euler (ZXZ) convention
+
+To initaially instanciate the Tensor Class, following needs to be inputted in python:
+
+.. code-block:: python
+
+    import numpy as np
+    from santex.tensor import Tensor
+
+    tensor = Tensor()
+
+Let's load some atiffness matrix values for forsterite written in voigt notation. The values can be seen in voigt matrix format
+
+.. code-block:: python
+
+    cij_forsterite = np.array([[320.5,  68.15,  71.6,     0,     0,     0],
+    [ 68.15,  196.5,  76.8,     0,     0,     0],
+    [  71.6,   76.8, 233.5,     0,     0,     0],
+    [   0,      0,      0,     64,     0,     0],
+    [   0,      0,      0,      0,    77,     0],
+    [   0,      0,      0,      0,     0,  78.7]])
+
+To convert this voigt notation to tensor notation, a user can call ``voigt_to_tensor()`` method within Tensor class as follows:
+
+.. code-block:: python
+    cijkl_forsterite = tensor.voigt_to_tensor(cij_forsterite)
+
+This gives a 3*3*3*3 array with the converted voigt in full tensor notation as seen
+
+.. code-block:: python
+
+    array([[[[320.5 ,   0.  ,   0.  ],
+         [  0.  ,  68.15,   0.  ],
+         [  0.  ,   0.  ,  71.6 ]],
+
+        [[  0.  ,  78.7 ,   0.  ],
+         [ 78.7 ,   0.  ,   0.  ],
+         [  0.  ,   0.  ,   0.  ]],
+
+        [[  0.  ,   0.  ,  77.  ],
+         [  0.  ,   0.  ,   0.  ],
+         [ 77.  ,   0.  ,   0.  ]]],
+
+
+       [[[  0.  ,  78.7 ,   0.  ],
+         [ 78.7 ,   0.  ,   0.  ],
+         [  0.  ,   0.  ,   0.  ]],
+
+        [[ 68.15,   0.  ,   0.  ],
+         [  0.  , 196.5 ,   0.  ],
+         [  0.  ,   0.  ,  76.8 ]],
+
+        [[  0.  ,   0.  ,   0.  ],
+         [  0.  ,   0.  ,  64.  ],
+         [  0.  ,  64.  ,   0.  ]]],
+
+
+       [[[  0.  ,   0.  ,  77.  ],
+         [  0.  ,   0.  ,   0.  ],
+         [ 77.  ,   0.  ,   0.  ]],
+
+        [[  0.  ,   0.  ,   0.  ],
+         [  0.  ,   0.  ,  64.  ],
+         [  0.  ,  64.  ,   0.  ]],
+
+        [[ 71.6 ,   0.  ,   0.  ],
+         [  0.  ,  76.8 ,   0.  ],
+         [  0.  ,   0.  , 233.5 ]]]])
+
+The converted tensor can be converted back to voigt notation as:
+
+.. code-block:: python
+    cij_forsterite = tensor.tensor_to_voigt(cijkl_forsterite)
+
+Which then returns the 6*6 voigt notation of the tensor
+
+.. code-block:: python
+
[…]
