### 2024-09-24 #6886 @utpal-singh
Hi @elena-pascal ,

Thank you for elaborating the issue earlier, I have included the documentation which is live at [https://santex.readthedocs.io/en/latest/](https://santex.readthedocs.io/en/latest/), please let me know if its accessible now and easy to understand.
