+++ docs/source/workingOnSantex.rst
@@ -217,25 +217,37 @@ Compare original and clean datasets
 .. code-block:: python
 
     ebsdfile.plot()
-    ebsdfile.plot(df = filtered_df_grain_boundary)
+    ebsdfile.plot(data = filtered_df_grain_boundary)
+
+The plot method of the EBSD class can take in a few parameters namely,
+
+    data (pandas.DataFrame, optional): DataFrame containing EBSD data. If not provided, uses stored data.
+    rotation_angle (int, optional): Angle by which to rotate the EBSD data (in degrees). Accepts 0, 90, 180, 270.
+    inside_plane (bool, optional): If True, rotates the EBSD data inside the plane. If False, rotates outside the plane. Default is True.
+    mirror (bool, optional): If True, mirrors the EBSD data horizontally before rotating. Default is False.
+    save_image (bool, optional): If True, saves the plot as an image. Default is False.
+    image_filename (str, optional): Filename to use when saving the image. Required if save_image is True.
+    dpi (int, optional): Dots per inch for the saved image. Default is 300.
+    cmap (str, optional): Colormap to use for plotting. Default is 'viridis'.
+    legend_location (str, optional): Location of the legend. Options are 'upper right', 'upper left', 'lower right', 'lower left'. Default is 'upper right'.
 
 
-Calculating Anisotropy from EBSD file:
+Calculating Anisotropy from EBSD file
 ==============
 
-To prepare the dataframe, the user can instanciate Material class as follows:
+To prepare the dataframe, the user can instanciate Material class as follows, be sure to check if the Material class is imported via ``from santex.material import Material``:
 
 .. code-block:: python
 
     material_instance = Material()
 
-The densities are within the santex package registry, and can be accessed via the following directive. The pressure in GPa and Temperature can be entered as:
+The densities are within the santex package registry, and can be accessed via the following directive. The pressure in GPa (here example: 2GPa) and Temperature (here 1500 degrees Celsius) can be entered as:
 
 .. code-block:: python
 
-    rho_Fo = material_instance.load_density("Forsterite", 2, 1500)
-    rho_diop = material_instance.load_density("Diopside", 2, 1500)
-    rho_ens = material_instance.load_density("Enstatite", 2, 1500)
+    rho_Fo = material_instance.load_density("Forsterite", pressure = 2, temperature = 1500)
+    rho_diop = material_instance.load_density("Diopside", pressure = 2, temperature = 1500)
+    rho_ens = material_instance.load_density("Enstatite", pressure = 2, temperature = 1500)
 
 The registry for elastic tensors within santex can be accessed via the following commands. The pressure and temperacture
 conditions can be entered via the keyword PRESSURE and TEMP
@@ -255,6 +267,8 @@ For preparing dataframes for seismic anisotropy, the elasticstiffness tensors an
 
 The euler angles for the phases should also be bundled in list as:
 
+**Note:** Make sure to import Anisotropy class first using ``from santex.anisotropy import Anisotropy``
+
 .. code-block:: python
 
     forsterite = ebsdfile.get_euler_angles(phase = 1, data=filtered_df_grain_boundary)
@@ -263,6 +277,8 @@ The euler angles for the phases should also be bundled in list as:
     euler_angles = [forsterite, enstatite, diopside]
 
 The anisotropy can then be calculated as following:
+
+
 
 .. code-block:: python
 
@@ -275,11 +291,23 @@ To look at the plots for the seismic velocities, following command can be entere
 
     anis.plot()
 
[…]
