+++ README.md
@@ -27,3 +27,11 @@ or
 ```bash
 pip install santex
 ```
+
+# Workflow
+
+<img src="./santex.png">
+
+# Contacts
+
+| **Utpal Singh** | utpal.singh@sydney.edu.au
