### 2024-10-04 fe9cd36
Update setup.py for platforms
