### 2024-10-08 5fcfb84
Update README.md for istallation
