### 2024-10-21 utpal-singh/SAnTex#7 @elena-pascal
JOSS REVIEW -- slightly more general install requirements
This relates to https://github.com/openjournals/joss-reviews/issues/6886

This is a super easy one. It is great that you identified different requirements for different operating systems. I tried to follow your exact steps and failed because the way python is setup on my work PC is very convoluted but it makes it a good case study.
What I needed to know before I followed the steps in the README is:

- [ ] minimum Python version expected. I suppose 3.9
- [ ] Windows requires a C compiler (which can be install with conda for instance as you explain)
