### 2024-11-03 #6886 @utpal-singh
> (emailed @SunilAnandatheertha, will find replacement reviewer if I don't hear back soon)

Hi @rkurchin, Thank you for taking the time to find additional reviewers. Given that the review process has already extended over five months and we’re still seeking another reviewer, I’m wondering if there might be an option to proceed with just one. I ask because I have other work based on the SAnTex library that I’m hoping to publish and present at upcoming conference, and the timeline is a little challenging. Could you let me know if moving forward with a single reviewer might be possible in this case?
