### 2024-11-03 utpal-singh/SAnTex#7 @utpal-singh
Hi @elena-pascal, thank you for reaching out, and thanks a lot for the suggestion, **were you finally able to install it in your system?** I'll be including more general instructions for the installation, thank you for pointing out. The installation is very easy for linux or mac systems, for window's it's a little work for now
