### 2024-11-05 #6886 @rkurchin
👋 @lfgmorales, @kilir, and/or @benholtzman, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Because this is a replacement for a nonresponsive reviewer, we would appreciate an expedient review if at all possible, though of course understand that everyone has many demands on their time.
