### 2024-11-06 utpal-singh/SAnTex#8 @elena-pascal
JOSS REVIEW -- motivation for rewriting EBSD tools
This relates to https://github.com/openjournals/joss-reviews/issues/6886

I am slowly going through the code.

One thing that is not clear to me and the paper is not helping either, is the need for step 1.

> 1. Processing of EBSD data: Facilitates the processing and cleaning of EBSD data,
calculating orientation distribution functions, pole density functions, and inverse pole
figures(Johnstone et al., 2020).

What does `SAnTex` add that `orix` does not offer. The paper should be more clear about what functionalities  `SAnTex` wraps over and what functionalities it adds and more importantly why.

For instance, I see you wrote a lot of orientation tools but you also use `orix` orientation tools. The code is not clear here either; instead of inheriting and extending `orix` objects there seems to be new objects introduced that are meant to interact with `orix` objects.
