### 2024-11-06 utpal-singh/SAnTex#9 @elena-pascal
JOSS REVIEW -- evidence that `Processing EBSD data` yields correct results
This relates to https://github.com/openjournals/joss-reviews/issues/6886

This is still about `SAnTex` first claimed functionality:

> Processing of EBSD data: Facilitates the processing and cleaning of EBSD data,
> calculating orientation distribution functions, pole density functions, and inverse pole
> figures(Johnstone et al., 2020).
>

Because you wrote your own orientation tools, I would expect to see both tests that check the maths does what you expect as well as results that when applied on real data it produces similar results to literature.
