### 2024-11-06 utpal-singh/SAnTex#11 @elena-pascal
JOSS REVIEW - tests don't cover all code functionalities promised in the paper
This relates to https://github.com/openjournals/joss-reviews/issues/6886

Since you differentiate between 5 functionalities the code can offer, I would expect tests that cover at least parts of each of these.
