### 2024-11-06 utpal-singh/SAnTex#12 @elena-pascal
JOSS REVIEW -- explanation of where the data in tests comes from missing
This relates to https://github.com/openjournals/joss-reviews/issues/6886

There should be some information of what the tests are testing against. Because if your data comes from previously running the function, that means something else to the scenario where you use real data or literature data to test against.
