### 2024-11-06 utpal-singh/SAnTex#13 @elena-pascal
JOSS REVIEW -- no data examples in the paper
This relates to https://github.com/openjournals/joss-reviews/issues/6886

You have a lot of beautiful data in the notebooks. You should use some of it to show the software functionality in the paper. And I don't mean show as in show the images, I mean show that the software produces physical results by comparing with literature.
