### 2024-11-06 utpal-singh/SAnTex#14 @elena-pascal
JOSS REVIEW -- lack of proof for mathematical approach
This relates to https://github.com/openjournals/joss-reviews/issues/6886

The paper does a good job of emphasising that temperature and pressure are important variables to consider when computing strain. And it proposed eq. 2 to address that. But it does nothing to convince me it is sufficient. Especially because you say we have first principles calculations are used in literature for taking into account these variables. But if we can Taylor expand to approximate their effect why bother with first principles.

I think because you have such a strong set up, you now need to bring very strong proof for your approach. What I would reasonably expect is a link to a paper where you show that your second order Taylor expansion as you implement in `SAnTex` gives reasonable results. If that is not possible, a link to a paper where this approach is used and yield satisfactory results. If that is not possible, some theoretical paper that shows this is a reasonable approach to pursue.
