### 2024-11-06 utpal-singh/SAnTex#15 @elena-pascal
JOSS REVIEW -- implementation steps unclear
This relates to https://github.com/openjournals/joss-reviews/issues/6886

`Method` section in the paper includes what could have been useful implementation conceptual steps. But in the current form they are very vague.

- [ ] what is `effective tensor`.
- [ ] how do you calculate the constants given the equations above
- [ ] what is a `reference stiffness tensor`
- [ ] what are `Voigt-Reuss-Hill bounds`
