### 2024-11-13 #6886 @lfgmorales
Hi Rachel,

Thanks for the email

I am not entirely sure how to proceed here and what do we have to do, there was also an answer from Elena Pascal last week regarding this paper. Or this is not a paper “per se”, but just the code?

If you can clarify that, would be very helpful

all the best

Luiz

Luiz F. G. Morales
ETH Zürich
Structural Geology and Tectonics Group
Scientific Center for Optical and Electron Microscopy (ScopeM)
HPM C 57.3
Otto-Stern-Weg 3
8093 Zurich, Switzerland
Phone +41 44 633 37 46
Email: ***@***.***

> On 13 Nov 2024, at 16:29, Rachel Kurchin ***@***.***> wrote:
>
>
> (Emailed @lfgmorales <https://github.com/lfgmorales>, @kilir <https://github.com/kilir>, and @benholtzman <https://github.com/benholtzman>)
>
