### 2024-11-18 #6886 @benholtzman
Hello-- Sorry I hadn't responded sooner.
I don't have the capacity now to test this package (and haven't worked on
these problems for a long time),
but it seems very similar to MTEX, so I suggest any of the people on this
page:
https://mtex-toolbox.github.io/people
Best wishes,
Ben

On Mon, Nov 18, 2024 at 11:58 AM Elena Pascal ***@***.***>
wrote:

> Can I suggest @hakonanes <https://github.com/hakonanes> ? I don't know if
> he will be available but he is an EBSD person and pretty active on github.
>
