### 2024-11-22 #6886 @utpal-singh
> Hello @utpal-singh, apologies for taking so long to get to this, work has been silly.
>
> I am looking through the code and your notebooks. I think at this point, the paper and code is missing a way to validate its functionality. See the PRs. I cannot check the that you have automated tests that confirm the functionality of the code, since your tests only cover one or two (not clear) of the 5 functionalities. It would also be nice to have some information of where the values you assert against come from in tests. Especially because you have a lot of info in your notebooks, which are at the moment hidden, not really sure why they are part of this repo since you don't mention any of it in the paper.
>
> Feel free to present the code and its results at the conference, and feel free to mention it is under review with JOSS. I don't see any conflict, or am I missing something?

Hi @elena-pascal, apologies for the delayed response, I have been attending a conference over the past few days. I appreciate your thorough feedback, and your concerns regarding testing and validation of the code. Thank you for your valuable insights and I look forward to address the review points and the elaborated reviews. Although it might be a bit slow, since I am also going for a leave, but hopefully I would be able to work on the reviews.
