### 2025-01-27 #6886 @hakonanes
Hi @utpal-singh, and happy new year to you too!

I've read the paper and gone through the [*Working with SAnTex* tutorial](https://santex.readthedocs.io/en/latest/workingOnSantex.html) in the docs. Taken together, they show promise in demonstrating an open and reproducible workflow for calculation of seismic wave velocities and anisotropy. I particularly like the flexibility with which a workflow that suits the user can be constructed.

However, the paper lacks a compelling demonstration of the software's claimed functionality. And the software needs to communicate more clearly what is done in each operation (better documentation, visuals). I'll list comments to the paper here, and report questions and suggestions for the software in the SAnTex issue tracker.

Comments by line:
- L24: Reduce number of CPO references to one. For example a text book. There are too many references in the *Statement of need* for my liking.
- L27: Remove the second reference to *Mainprice and Nicolas (1989)*, cited just above.
- L33: A brief comparison and a reference to an existing software which provides "standard practice" analysis is needed. Seems to me like MTEX is such a software (https://mtex-toolbox.github.io/CPOSeismicProperties.html).
- L33-35: Unclear: "... use stiffness tensors corresponding to room temperature and pressure conditions or consider only the effect of pressure.". Corresponding to room temperature and *room pressure conditions*? I don't understand which pressure conditions are ment here.
- L52: I recommend a text book reference for the equations (1)-(3).
- Eq. (2): A one-sentence description of how the pressure and temperature affects the tensor would be useful. Drop the double "::" before the equation. Fix formatting of subscripts.
- Eq. (3): Split onto two lines. Fix formatting of subscripts.
- L63: Briefly discuss the validity of the implementation. An example is needed, preferably with visuals.
- L67: Re-formulate "more functionalities" to something more concrete.
- L74: Briefly explain what Voigt-Reuss-Hill bounds are.
- L78: Rename to "Features"?
- Figure 1: Drop the API calls, explain the workflow in words with some nice schematics instead. The API in the figure is *already* outdated (`santex.Material` is now available from `santex.material.Material`). The API is bounds to change further in the future, to meet different users' needs, so I'd refrain from referencing API calls completely. This is a good place to rather refer to the API in the docs.
- L80: (Disclaimer, I'm one of the orix maintainers). Make it clear that pole figures and inverse pole figures are produced with orix. There's a reference, good. But, the way it reads now, the reference seems to point to somewhere you read more about these data representations, rather than the software that provides this functionality.
- L83: Unclear formulation: tensor operations and rotations using Euler angles?
- L89: No need to capitalize "Seismic Anisotropy". Also, I suggest to drop the list format of "Functionality: functionality in slightly higher detail", and just keep what's after the colons.
- L91: Don't just mention visualizations, show them in the paper!
- L92: Format variables in italic with subscripts.
