### 2025-04-23 #6886 @elena-pascal
Hello @utpal-singh, I went through the issues I added and closed them as you very answered all of them with great care. Thank you for that, looking good.

I just have one more checkpoints to ticks, the community guidelines.

Additionally, I still think you would benefit from adding a reference to your notebooks in the README.

Feel free to have a look at the other submissions READMEs. Almost there!
