### 2025-04-23 #6886 @hakonanes
Thank you for accomodating all my comments, @utpal-singh. I find the paper to read much better now.

I have one remaining comment: Under "SAnTex Analysis / EBSD data Analysis" in Figure 2, you write that SAnTex can "Compute orientation distribution function". As far as I can tell, SAnTex can provide the plot pole densities, but not calculate an ODF. If I'm right, I'd remove that line.
