### 2025-04-23 #6886 @utpal-singh
Hi @hakonanes, Thank you for taking the time to review the submission, and thanks for pointing this out. I’ve removed that line. It was included when we were writing with ODF calculations in mind, but that functionality isn’t part of this version at the moment.
