### 2025-04-26 #6886 @hakonanes
Thanks. I think the paper showcases the software's functionality nicely. Users curious to learn more and use the software will find a rich documentation online. Granted my suggested change is part of the final proof, I'm happy for this submission to be accepted!
