### 2025-04-27 #6886 @rkurchin
@utpal-singh, it looks like we're ready to finalize this submission for JOSS! In addition to the last reviewer request above, please send me the information requested in the comment above (version tag and DOI) and check the other things as well. I will send along some editorial comments shortly...
