### 2025-04-28 #6886 @elena-pascal
Thank you for being so diligent with implementing our comments. The paper is now an excellent introduction to the functionality of the software and the repo includes all tools needed for a user to reproduce the results, contribute and raise issues. I think it makes a great contribution to the scientific open software landscape.
