### 2025-04-28 #6886 @rkurchin
Editorial remarks...
- sentence spanning lines 11-13 has some singular-plural disagreement from the antecedent clause: I would change "influence" to "influences"
- 17: I presume Vp, Vs are meant to be subscripted? (likewise Vbulk later, and also in citation on line 214)
- 39: these offsetting dashes should be em-dashes – like that one there.
- 55: hyphenated "4th-order" and also "moduli" has only one "i" in it 😉
- equation between lines 65 and 66 (and later on): consider making $p$ and $P_0$ either both upper-case or both lowercase? If the way you've rendered it is a strong convention in your subfield, then ignore this.
- same equation (and also later on): I would put parentheses around the $\Delta p$ and $\Delta T$ in the higher-order term for clarity, i.e. $(\Delta T)^2$
- line 67-68: remove the ~ in the ambient condition specification – if this is meant to indicate approximation then change to a `\simeq`
- 71: Decapitalize "Temperature", change "has" to "have", and add a "the" before "stiffness tensor"
- 74: either add an "a" before "homogeneous" or change to "...as a homogeneously distributed isotropic phase..." (I lean towards the second option)
- equation between 75 and 76: Is $f_{melt}$ the same as $f$? I presume yes from context – you should indicate that consistently and I would also remove the outer parentheses in the last term of the RHS, because they suggest that $f_{melt}$ is a function, when it is just some scalar multiplier.
- Same equation: Do `\mathrm{melt}` for the subscript, since $melt$ is not a variable
- 99: add a "that" after "demonstrate"
- 100: add space between "Figure" and "1c" and period after "vs"
- throughout: be consistent whether figure references are "Figure" or "Fig" or "Fig."
-  Figure 1 caption: add space before (c) on first line, decapitalize "Bounds" on second-to-last line, add a period at the end of the caption
