### 2025-05-22 #6886 @utpal-singh
Hi @rkurchin,
Thank you very much for your time and effort throughout the editorial process!
And many thanks to @elena-pascal and @hakonanes for your insightful and constructive reviews.

@rkurchin, I’ve just double-checked the Zenodo record and the Santex repository — everything looks good on my end. I believe we’re ready to proceed with the acceptance.
