### 2025-05-23 #6886 @kthyng
@utpal-singh your paper has gained some length — what do you think about moving some of it to your docs to preserve the short paper length?
