### 2025-05-27 #6886 @utpal-singh
Hi @kthyng, Thank you for your message. we would prefer to retain the current length of the paper, as it reflects the comprehensive revisions made through the review process. However, I understand the need for reducing length, and if shortening is absolutely necessary, I’m happy to explore ways to move appropriate sections to supplementary materials or documentation. Please let me know your preference.
